Add Windows Server 2022 Desktop Packer templates
This was originally based on https://github.com/StefanScherer/packer-windows/blob/main/windows_2022.json, then heavily modified.

The build is split up into two separate Packer builds: an installation build, and a provisioning build which works on a copy-on-write image backed by the installation build's output. This is done so that working on the provisioning step doesn't require you to complete a full reinstallation of Windows first.